--- config/default.json
@@ -1,3 +1,0 @@
-{
-  "name": "My Express App"
-}
