--- config/development.json
@@ -1,6 +1,0 @@
-{
-  "name": "My Express App - Development",
-  "mail": {
-    "host": "dev-mail-server"
-  }
-}
