--- config/production.json
@@ -1,6 +1,0 @@
-{
-  "name": "My Express App - Production",
-  "mail": {
-    "host": "prod-mail-server"
-  }
-}
