+++ index.js
@@ -1,108 +1,14 @@
 
-const config = require('config')
-const helmet = require('helmet')
 const Joi = require('joi')
 const express = require('express');
 const app = express();
-const logger = require('./logger')
-const morgan = require('morgan')
-
+const genres = require('./routes/genres')
 
 app.use(express.json())
-app.use(express.urlencoded({ extended: true }))
-// app.use(logger); 
-app.use(helmet())
+app.use('/api/genres', genres);
 
 
-//COnfiguration
-console.log('Application Name: ')
 
-
-if(app.get('env') === 'development') {
-  app.use(morgan('tiny'))
-  console.log('Morgan is enabled')
-}
-
-
-const genres = [
-  {
-    id: 1,
-    name: 'action',
-  },
-  {
-    id: 2,
-    name: 'horror',
-  },
-  {
-    id: 3,
-    name: 'romance',
-  },
-  {
-    id: 4,
-    name: 'drama',
-  },
-  {
-    id: 5,
-    name: 'comedy',
-  },
-  {
-    id: 6,
-    name: 'documentary',
-  },
-];
-
-function validateGenre(genre) {
-  const schema = Joi.object().keys({
-    name: Joi.string().min(3).required()
-  })
-
-  return schema.validate(genre)
-}
-
-app.get('/api/genres', (req, res) => {
-  res.send(genres);
-});
-
-app.post('/api/genres', (req, res) => {
-  const genre = {
-    id: genres.length + 1,
-    name: req.body.name,
-  };
-  genres.push(genre);
-  res.send(genre);
-});
-
-
-app.get('/api/genres/:id', (req, res) => {
-  const genre = genres.find((g) => g.id === parseInt(req.params.id))
-  if (!genre) {
-    return res.status(404).send('There is no genre with the given ID')
-  }
-  res.send(genre)
-} )
-
-app.put('/api/genres/:id', (req, res) => {
-  let genre = genres.find((g) => g.id === parseInt(req.params.id))
-  if(!genre) {
-    return res.status(404).send('There is no genre found with the given ID')
-  }
-  const { error }= validateGenre(req.body)
-  if(error) {
-    return res.status(400).send(error.details[0].message)
-  }
-  genre.name = req.body.name
-  res.send(genre)
-})
- 
-app.delete('/api/genres/:id', (req, res) => {
-    let genre = genres.find((g) => g.id === parseInt(req.params.id));
-    if (!genre) {
-      return res.status(404).send('There is no genre found with the given ID');
-    }
-    const index = genres.indexOf(genre)
-    genres.splice(index, 1)
-    res.send(genre)
-})
