--- logger.js
@@ -1,7 +1,0 @@
-
-function log(req, res, next) {
-  console.log('Logging...')
-  next()
-}
-
-module.export = log;
