+++ package.json
@@ -1,9 +1,6 @@
 {
   "dependencies": {
-    "config": "^3.3.6",
     "express": "^4.17.1",
-    "helmet": "^4.4.1",
-    "joi": "^17.4.0",
-    "morgan": "^1.10.0"
+    "joi": "^17.4.0"
   }
 }
