+++ routes/genres.js
@@ -1,0 +1,85 @@
+
+const express = require('express')
+const router = express.Router();
+
+const genres = [
+  {
+    id: 1,
+    name: 'action',
+  },
+  {
+    id: 2,
+    name: 'horror',
+  },
+  {
+    id: 3,
+    name: 'romance',
+  },
+  {
+    id: 4,
+    name: 'drama',
+  },
+  {
+    id: 5,
+    name: 'comedy',
+  },
+  {
+    id: 6,
+    name: 'documentary',
+  },
+];
+
+function validateGenre(genre) {
+  const schema = Joi.object().keys({
+    name: Joi.string().min(3).required()
+  })
+
+  return schema.validate(genre)
+}
+
+router.get('/', (req, res) => {
+  res.send(genres);
+});
+
+router.post('/', (req, res) => {
+  const genre = {
+    id: genres.length + 1,
+    name: req.body.name,
+  };
+  genres.push(genre);
+  res.send(genre);
+});
+
+
+router.get('/:id', (req, res) => {
+  const genre = genres.find((g) => g.id === parseInt(req.params.id))
+  if (!genre) {
+    return res.status(404).send('There is no genre with the given ID')
+  }
+  res.send(genre)
+} )
+
+router.put('/:id', (req, res) => {
+  let genre = genres.find((g) => g.id === parseInt(req.params.id))
+  if(!genre) {
+    return res.status(404).send('There is no genre found with the given ID')
+  }
+  const { error }= validateGenre(req.body)
+  if(error) {
+    return res.status(400).send(error.details[0].message)
+  }
+  genre.name = req.body.name
+  res.send(genre)
+})
+ 
+router.delete('/:id', (req, res) => {
+    let genre = genres.find((g) => g.id === parseInt(req.params.id));
+    if (!genre) {
+      return res.status(404).send('There is no genre found with the given ID');
+    }
+    const index = genres.indexOf(genre)
+    genres.splice(index, 1)
+    res.send(genre)
+})
+
+module.exports = router;
